requirements.txt
Django==2.1.*
requests==2.19.*
google-auth==1.5.*
